requirements.txt
anyio==3.6.2
AnyQt==0.2.0
arff==0.9
async-generator==1.10
backcall==0.2.0
baycomp==1.0.2
Bottleneck==1.3.6
CacheControl==0.12.11
certifi==2022.12.7
cffi==1.15.1
chardet==5.0.0
charset-normalizer==2.0.12
commonmark==0.9.1
contextvars==2.4
cryptography==39.0.1
cycler==0.11.0
dataclasses==0.8
decorator==4.4.2
dictdiffer==0.9.0
docutils==0.18.1
entrypoints==0.4
et-xmlfile==1.1.0
h11==0.12.0
httpcore==0.13.7
httpx==0.19.0
idna==3.4
immutables==0.19
importlib-metadata==4.8.3
ipykernel==5.5.6
ipython==7.16.3
ipython-genutils==0.2.0
jedi==0.17.2
jeepney==0.7.1
joblib==1.1.1
jupyter-client==7.1.2
jupyter-core==4.9.2
keyring==23.4.1
keyrings.alt==4.1.0
kiwisolver==1.3.1
lockfile==0.12.2
matplotlib==3.3.4
msgpack==1.0.4
nest-asyncio==1.5.6
networkx==2.5.1
numpy==1.19.5
openpyxl==3.1.1
openTSNE==0.7.1
orange-canvas-core==0.1.29
orange-widget-base==4.19.0
Orange3==3.31.1
packaging==21.3
pandas==1.1.5
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
pkg_resources==0.0.0
prompt-toolkit==3.0.36
ptyprocess==0.7.0
pycparser==2.21
Pygments==2.14.0
pyparsing==3.0.9
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.1
pyqtgraph==0.11.1
PyQtWebEngine==5.15.5
PyQtWebEngine-Qt5==5.15.2
python-dateutil==2.8.2
python-louvain==0.16
pytz==2022.7.1
PyYAML==6.0
pyzmq==25.0.0
qasync==0.23.0
qtconsole==5.2.2
QtPy==2.0.1
requests==2.27.1
rfc3986==1.5.0
scikit-learn==0.24.2
scikit-multilearn==0.2.0
scipy==1.5.4
SecretStorage==3.3.3
serverfiles==0.3.1
six==1.16.0
sklearn==0.0.post1
sniffio==1.2.0
threadpoolctl==3.1.0
tornado==6.1
traitlets==4.3.3
typing_extensions==4.1.1
urllib3==1.26.14
wcwidth==0.2.6
xlrd==2.0.1
XlsxWriter==3.0.8
zipp==3.6.0
